Fix drag behavior when page is empty
